:memo: Feat : Change 이미지 업로드 로직 수정

+++ sever/controllers/users/update.js
@@ -29,30 +29,57 @@ module.exports = {
       //만약 파일이 기존의 이미지와 다르다면, 기존의 파일을 삭제하고 새이미지
       //어차피 사진은 기존의 사진이 들어가도 multer에서 시간 순으로 바꿨기 때문에
       //기존 사진 지워도 됨
+      
+      if (user.profile_img != new_path) {
         const original_profile_img = user.profile_img.slice(34);
-      if (fs.existsSync(path.join(__dirname,"..","..","profile_img",`${original_profile_img}`))) {
-        fs.unlinkSync(path.join(__dirname,"..","..","profile_img",`${original_profile_img}`));
+        if (fs.existsSync(path.join(__dirname,"..","..","profile_img",`${original_profile_img}`))) {
+          fs.unlinkSync(path.join(__dirname,"..","..","profile_img",`${original_profile_img}`));
+        }
+  
+        await User.update(
+          {
+            nickname: nickname,
+            phone_number: phone_number,
+            profile_img: new_path,
+            height: height,
+            weight: weight,
+            gender: gender,
+            sns_url: sns_url,
+          },
+          { where: { nickname: req.session.user.nickname } }
+          );
+          const newUserInfo = await User.findOne({ where: { nickname: nickname } });
+          return res.status(200).json({
+              message: "변경 되었습니다.",
+              data: {
+                  newUserInfo
+              }
+          })
+        
       }
+      //사진 정보가 안바뀌었을 때 
+      else {
+        await User.update(
+          {
+            nickname: nickname,
+            phone_number: phone_number,
+            height: height,
+            weight: weight,
+            gender: gender,
+            sns_url: sns_url,
+          },
+          { where: { nickname: req.session.user.nickname } }
+          );
+          const newUserInfo = await User.findOne({ where: { nickname: nickname } });
+          return res.status(200).json({
+              message: "변경 되었습니다.",
+              data: {
+                  newUserInfo
+              }
+          })
 
-      await User.update(
-        {
-          nickname: nickname,
-          phone_number: phone_number,
-          profile_img: new_path,
-          height: height,
-          weight: weight,
-          gender: gender,
-          sns_url: sns_url,
-        },
-        { where: { nickname: req.session.user.nickname } }
-        );
-        const newUserInfo = await User.findOne({ where: { nickname: nickname } });
-        return res.status(200).json({
-            message: "변경 되었습니다.",
-            data: {
-                newUserInfo
-            }
-        })
+      }
+     
     
     }
